Split definition into XMLString.h
